+++ src/app/cpp/OutlineModels.hpp
@@ -50,6 +50,40 @@ struct OutlineItem {
     friend bool operator==(const OutlineItem&, const OutlineItem&) = default;
 };
 
+struct TrashItem {
+    QString title;
+    bool wholeSection{};
+    bool restorable{};
+
+    friend bool operator==(const TrashItem&, const TrashItem&) = default;
+};
+
+class TrashListModel : public QAbstractListModel {
+    Q_OBJECT
+    QML_ELEMENT
+    QML_UNCREATABLE("Provided by NotebookViewModel")
+    Q_PROPERTY(int count READ rowCount NOTIFY countChanged FINAL)
+
+public:
+    static constexpr int kTitleRole = Qt::UserRole + 1;
+    static constexpr int kSectionRole = Qt::UserRole + 2;
+    static constexpr int kRestorableRole = Qt::UserRole + 3;
+
+    explicit TrashListModel(QObject* parent = nullptr);
+
+    [[nodiscard]] int rowCount(const QModelIndex& parent = {}) const override;
+    [[nodiscard]] QVariant data(const QModelIndex& index, int role) const override;
+    [[nodiscard]] QHash<int, QByteArray> roleNames() const override;
+
+    void setItems(std::vector<TrashItem> items);
+
+signals:
+    void countChanged();
+
+private:
+    std::vector<TrashItem> m_items;
+};
+
 class OutlineListModel : public QAbstractListModel {
     Q_OBJECT
     QML_ELEMENT